### 2022-06-01 #4409 @danielskatz
👋 @zpace & @bmorris3 - Would either or both of you be willing to review this submission for JOSS?
