### 2022-06-06 #4409 @danielskatz
👋 @nespinoza & @hayesla - Would either or both of you be willing to review this submission for JOSS?
