### 2022-06-09 #4409 @ygrange
Oh that actually looks interesting, I have done some abundance stuff in the past so I could pick it up. But I have to say that I will probably have zero time the coming two weeks, so at earliest I could start looking at this in three weeks.
