### 2022-06-09 #4461 @editorialbot
[REVIEW]: Starmatrix: Modelling nucleosynthesis of galactic chemical elements
**Submitting author:** @xuanxu (<a href="http://orcid.org/0000-0001-7699-3983">Juanjo Bazán</a>)
**Repository:** https://github.com/xuanxu/starmatrix
**Branch with paper.md** (empty if default branch):
**Version:** v1.7.0
