### 2022-06-27 #4461 @ygrange
Let me start with the first pass of very minor comments, because I started by reading the README, CONTRIBUTING and CODE_OF_CONDUCT md files. For completeness: none of those are blockers.

### README file:

*  I assume `pip install` installs the most recent release version. I would suggest to make that explicit in the README

To me the fact that the "Edge" header is so far below the "Installation" one is confusing. If you'd ask me I'd advice for the following order:

* Installation
* Edge
* Test installation
* Test suite
* (and then the rest as it is)

Minor detail: the layout of the references is different between the two tables. It may be neater to have the same format in both. Also there are some minor inconsistencies within tables, depending on the answer of the editor above I may submit a pull request.

### CONTRIBUTING.md & CODE_OF_CONDUCT.md
You have pretty complete contributor guidelines. Few comments though:

* "Follow the code styles and practices already in place.", do you follow a documented standard? If so it would be useful to explicitly mention it, else you may consider giving some handles on what you use.
* You may want to explicitly mention the MIT license in the CONTRIBUTING.md (to save people from clicking back and forth)
* For completeness and clarity I would advice to link it from the README
