### 2022-06-27 #4461 @ygrange
Maybe one more for the README: could be useful that python >=3.7 is required (even though that should probably be obvious since 3.6 has been eol for a while).
