### 2022-06-27 75e7d6c
Minor fixes to README
Minor fixes to make it consistent.
