### 2022-06-28 #4461 @ygrange
Ok, I wanted to get into the more interesting part of the review, but I got some nasty hardware problems. So I will be a tad bit delayed while finding out how to do this on a temporary laptop (so I may have a bit less freedom wrt python versions and other stuff) :)
