### 2022-06-28 #4461 @ygrange
One more thing: The docs are linked from the README only via the docs - passing badge. I didn't realise that at first. I would strongly suggest adding a link to the docs somewhere in the text of the README too.
