### 2022-06-29 #4461 @ygrange
A random comment that I don't really want to put in my review but I noticed that if I make an overview of the test coverage, it shows to me that the parts of he code that are not covered are generally the last item of a list of parameter values to test. So for instance in supernovae.py:
```
def yields_from_leung_nomoto_2018_table8(feh):
   if feh <= -1.65:
    (...)
   elif 0.59 <= feh:
      return [0.0, 1.30e-3, 1.37e-10, 6.32e-10, 7.26e-2, 4.18e-4, 9.25e-4, 2.47e-1, 8.98e-2, 1.10e-2, 4.92e-1]
```

So it is quite often (I did not check alle the missing lines) it is basically the last elif statement that does not get tested. I am slightly puzzeled at this because your test coverage is great but it seems like this seems like a fairly simple way to make it even more complete.
