### 2022-06-29 xuanxu/starmatrix#54 @CFGrote
[JOSS Review] Installation
Here's my first review item for https://github.com/openjournals/joss-reviews/issues/4461

## Installation via `pip install starmatrix`
- no problem with the installation (in conda env).
- then run
```
starmatrix --generate-config
Created file: config-example.yml
```
=> OK

- Testing the installation as per README.md:
```
$> starmatrix --config config-example.yml

Running model with settings:

   z = 0.02
   sol_ab = as09
   imf = kroupa2002
   imf_alpha = 2.35
   imf_m_low = 0.15
   imf_m_up = 100
   m_min = 0.98
   m_max = 40.0
   binary_fraction = 0.15
   dtd_sn = rlp
   dtd_correction_factor = 1.0
   sn_yields = iwa1998
   output_dir = results
   matrix_headers = True
   return_fractions = False
   integration_step = logt
   deprecation_warnings = True
   expelled_elements_filename = ejections.dat
   total_time_steps = 300

Traceback (most recent call last):
  File "/opt/anaconda3/envs/starmatrix_review/bin/starmatrix", line 8, in <module>
    sys.exit(main())
  File "/opt/anaconda3/envs/starmatrix_review/lib/python3.10/site-packages/starmatrix/cli.py", line 41, in main
    model.Model(context).run()
  File "/opt/anaconda3/envs/starmatrix_review/lib/python3.10/site-packages/starmatrix/model.py", line 16, in __init__
    self.init_variables()
  File "/opt/anaconda3/envs/starmatrix_review/lib/python3.10/site-packages/starmatrix/model.py", line 21, in init_variables
    self.context["expelled"] = elements.Expelled(expelled_elements_filename=self.context["expelled_elements_filename"])
  File "/opt/anaconda3/envs/starmatrix_review/lib/python3.10/site-packages/starmatrix/elements.py", line 16, in __init__
    self.read_expelled_elements_file(expelled_elements_filename)
  File "/opt/anaconda3/envs/starmatrix_review/lib/python3.10/site-packages/starmatrix/elements.py", line 36, in read_expelled_elements_file
    expelled_data = open(filename, "r")
FileNotFoundError: [Errno 2] No such file or directory: 'ejections.dat'
```

## Installation from source
```
cd starmatrix
pip install -e .
```

- Generate config
```
$> starmatrix --generate-config
Created file: config-example.yml
```

- Test
```
$> starmatrix --config config-example.yml

Running model with settings:

   z = 0.02
   sol_ab = as09
   imf = kroupa2002
   imf_alpha = 2.35
   imf_m_low = 0.15
   imf_m_up = 100
   m_min = 0.8
   m_max = 100.0
   binary_fraction = 0.15
   dtd_sn = rlp
   dtd_correction_factor = 1.0
   sn_yields = iwa1998
   output_dir = results
   matrix_headers = True
   return_fractions = False
   integration_step = logt
   deprecation_warnings = True
   expelled_elements_filename = /home/grotec/Repositories/JOSSReviews/starmatrix/src/starmatrix/sample_input/expelled_elements
   total_time_steps = 300

Done. Output files ready in 'results' directory.
```
Have to remove the `config-example.yml` from the previous run though.

## Conclusion:
Installation from pip and installation from source result in different `config-example.yml` causing IO error in the former case.
