### 2022-06-29 xuanxu/starmatrix#54 @ygrange
I just came here to report the same thing. So basically replying to confirm this behaviour on my system :)
