### 2022-06-30 #4461 @ygrange
Few random comments I have when trying to check all the boxes of the checklist for which I won't open an issue:
1. One more for the README: I noticed the "guidelines" that it also is required to have an explicit mention of how to raise issues/ask questions. To me that would mean a single sentence in the readme "if you find a bug or have a question, please raise an issue on the issue board of this project." (potentially with a link)
2. Could you somewhere (either in the readme or readthedocs) add a sentence who your actual target audience is? Something as easy as "starmatrix is targeted at astronomers working on XYZ and need Q matrices to ABC". The goal of such a sentence would be for people who are doing something close to ABC to decided if they want to dive deeper into it.

I think all the comments I have given, including those two now cover the boxes I haven't checked yet :). I am pretty much done for now. Maybe I should end with what I would start with in a normal-style review:

I totally see the added value of startmatrix, it also seems like a very useful tool and I think I would have loved it existing during my PhD study. I enjoyed playing around with it (and probably will a bit pending the comments :) )
