### 2022-07-01 7551fd8
Mention default values in config-example.yml file
