### 2022-07-01 0cd7655
Mention license in the contributing guidelines
