### 2022-07-01 #4461 @xuanxu
> if I make an overview of the test coverage, it shows to me that the parts of he code that are not covered are generally the last item of a list of parameter values to test. So for instance in supernovae.py:
>
> ```
> def yields_from_leung_nomoto_2018_table8(feh):
>    if feh <= -1.65:
>     (...)
>    elif 0.59 <= feh:
>       return [0.0, 1.30e-3, 1.37e-10, 6.32e-10, 7.26e-2, 4.18e-4, 9.25e-4, 2.47e-1, 8.98e-2, 1.10e-2, 4.92e-1]
> ```
>
> So it is quite often (I did not check alle the missing lines) it is basically the last elif statement that does not get tested. I am slightly puzzeled at this because your test coverage is great but it seems like this seems like a fairly simple way to make it even more complete.

@ygrange: Yes, this bother me a little bit too, but those lines are not marked by Codecov as not tested (red), but as "partially tested" (yellow). I realized by that they mean that those lines are covered by tests but the coverage tool can't be certain every possible case is tested because there is not `else` clause.

A human eye can easily see that this `if` statatement covers all cases:
```python
if x < 5:
   # do something
elif x >=5:
  # do something else

```

But that coverage tool would mark the elif as "partial" just in case, because there is no default `else` branch. So we could get the coverage to 100% by rewriting it to:
```python
if x < 5:
   # do something
else:
  # do something else
```

In Starmatrix case I prefer to have the elif statements with the explicit values as in every case those are meaningful numbers obtained from literature and I prefer to have them visibly there (for instance see the [yellow lines in the DTDs file](https://app.codecov.io/gh/xuanxu/starmatrix/blob/main/src/starmatrix/dtds.py)). Does it make sense?
