### 2022-07-01 1713056
Define GCE
