### 2022-07-01 24b4332
Clarify what n.r. are
