+++ paper/paper.md
@@ -33,7 +33,7 @@ Using explicit (and configurable) values for *solar abundances*, *metallicity (z
 
 Calculating the contributions of chemical elements from different stars in several mass ranges is a necessary step for galactic chemical evolution models [@molla2017]. Usually this is done internally and the chemical yields are convoluted with a star formation history (SFH) inside the model, making it difficult to find reusable opensourced code for calculating the yield for a simple stellar population. The central module in Chempy [@rybizki2017] and the SYGMA module from the NUPYCEE framework [@sygma2018] are exceptions, but being part of bigger codebases they are not available as standalone libraries, not listed in the official Python Package Index and learning how to install and use them is not simple.
 
-`Starmatrix` isolates the SSP yield calculation step as a open source permissively licensed Python implementation, in a modular an flexible way that can be used to provide detailed datasets to be used as input for galactic chemical evolution models, to compare the validity of different yield sets from the literature or to assess different nucleosynthesis modeling assumptions [@bazan2003].
+`Starmatrix` isolates the SSP yield calculation step as an open source permissively licensed Python implementation, in a modular an flexible way that can be used to provide detailed datasets to be used as input for galactic chemical evolution models, to compare the validity of different yield sets from the literature or to assess different nucleosynthesis modeling assumptions [@bazan2003].
 
 # Features
 
@@ -42,7 +42,7 @@ Then, the time step is converted back into stellar mass intervals and a $Q$ matr
 
 All quantities in the $Q$ matrix are calculated on the base of the ejected mass from stars, using for it the dataset of yields per stellar mass that is entered as input for the code. The code is prepared to include any sets of stellar yields from the literature but if none is declared a default dataset for solar metallicity and with low and intermediate mass star yields from [@gavilan2006] and yields of massive stars from [@chieffi2004] will be used.
 
-![A sample plot using the output data from several `Starmatrix` runs. Lines show true yields `p` of Oxigen using different datasets for low and intermediate mass yields (columns) and for massive stars yields (rows), for a range of metallicity values (bottom horizontal axis) and different IMFs available as Starmatix' settings.\label{fig:oxigen}](sample_plot.png)
+![A sample plot using the output data from several `Starmatrix` runs. Lines show true yields `p` of Oxygen using different datasets for low and intermediate mass yields (columns) and for massive stars yields (rows), for a range of metallicity values (bottom horizontal axis) and different IMFs available as Starmatix' settings.\label{fig:oxygen}](sample_plot.png)
 
 Configurable parameters include metallicity, mass range for $Q$ matrices, fraction of binary systems, lower and upper mass limits for the IMF, total time steps for integration and yield correction factors. The code also includes a variety of options for `IMFs`, `solar abundances` data, Supernovae `yields` and `Delay Time Distributions` from the literature to choose from. A complete list of available options can be found in the documentation. To make the code more reusable and to allow it to be integrated in or called by other codebases, Starmatrix defines a base class for each of these parameters to be subclassed if needed so custom options can be easily added programmatically. The code also includes extensive test coverage.
 
