### 2022-07-01 #4461 @xuanxu
I've addressed most of the comments in the review here: https://github.com/xuanxu/starmatrix/pull/56
