+++ src/starmatrix/__init__.py
@@ -5,4 +5,4 @@ Starmatrix
 Modelling nucleosynthesis of galactic chemical elements
 
 """
-__version__ = "1.7.0"
+__version__ = "1.7.1"
