+++ requirements.txt
@@ -1,3 +1,3 @@
 pyyaml>=5.4.1
 numpy>=1.20
-scipy>=1.5
+scipy>=1.7
