### 2022-07-01 xuanxu/starmatrix#54 @xuanxu
Nice catch! The different `confing-example.yml` error was caused because I didn't add the changes to the latest release 🤦
That should be fixed now as I've released a new 1.7.1 version.
