+++ docs/configuration.rst
@@ -26,7 +26,7 @@ Starmatrix reads a configuration file where several input parameters (all of the
                                     # and Chieffi & Limongi, 2004, ApJ, 608, 405
         yield_corrections # A map of individual correction factors for items in the ejected data file
 
-Starmatrix will use its internal default values for all params for which no values are provided.
+Starmatrix will use its internal `default values`_ for all params for which no values are provided.
 
 If you want to use an existent configuration file as template for your own, you can run::
 
@@ -165,3 +165,30 @@ If you want to multiply all magnesium data by 2, and all Sulfur data by 0.87 you
         mg: 2
         s: 0.87
 
+
+Default values
+--------------
+
+If Starmatrix is run with an empty configuration file or using the file generated by the command ``starmatrix --generate-config`` with no modifications the set of default values that will be used for each parameter are the following:
+
+:z: 0.02
+:sol_ab: as09
+:imf: kroupa
+:imf_m_low: 0.15
+:imf_m_up: 100
+:m_min: 0.98
+:m_max: 40.0
+:total_time_steps: 300
+:binary_fraction: 0.15
+:dtd_sn: rpl
+:sn_yields: iwa1998
+:output_dir: results
+:matrix_headers: yes
+:return_fractions: False
+:integration_step: logt
+:dtd_correction_factor: 1.0 # No corrections
+:deprecation_warnings: True
+:expelled_elements_filename: data for z=0.02 from Gavilan et al, and Chieffi & Limongi
+:yield_corrections: # No corrections
+
+This set of values are a mix of solar values were possible and common options from the literature for the Milky Way for star masses from 0.98 to 40 Msun without correcting yields or SN factors.
