### 2022-07-06 3f24604
Expand information on modules and add links to code
