### 2022-07-08 fadbbba
Clarify SSP
