+++ README.rst
@@ -194,18 +194,18 @@ Supernovae yields
 
 The ``sn_yields`` param in the config file can be set to use any of the available supernova yields datasets from different papers/authors:
 
-:iwa1998: Data from Iwamoto, K. et al., 1999, ApJ 125, 439
-:sei2013: Data from Seitenzahl et al. 2013, MNRAS 429, 2, 1156–1172
-:ln2018-1: Data from Leung & Nomoto 2018, ApJ, Vol 861, Issue 2, Id 143, Tables 6/7
-:ln2018-2: Data from Leung & Nomoto 2018, ApJ, Vol 861, Issue 2, Id 143, Tables 8/9
-:ln2018-3: Data from Leung & Nomoto 2018, ApJ, Vol 861, Issue 2, Id 143, Tables 10/11
-:ln2020: Data from Leung & Nomoto 2020, ApJ, Vol 888, Issue 2, Id 80
-:br2019-1: Data from Bravo, E. et al., 2019, MNRAS, 482, Issue 4, 4346–4363, Table 3
-:br2019-2: Data from Bravo, E. et al., 2019, MNRAS, 482, Issue 4, 4346–4363, Table 4
-:gro2021-1: Data from Gronow, S. et al., 2021, A&A, Tables 3/A10 He+Core detonations
-:gro2021-2: Data from Gronow, S. et al., 2021, A&A, Tables 4/A8 He+Core detonations
-:mor2018-1: Data from Mori, K. et al., 2018, ApJ, 863:176 W7
-:mor2018-2: Data from Mori, K. et al., 2018, ApJ, 863:176 WDD2
+:iwa1998: Data from Iwamoto, K. et al., 1999
+:sei2013: Data from Seitenzahl et al. 2013
+:ln2018-1: Data from Leung & Nomoto 2018, Tables 6/7
+:ln2018-2: Data from Leung & Nomoto 2018, Tables 8/9
+:ln2018-3: Data from Leung & Nomoto 2018, Tables 10/11
+:ln2020: Data from Leung & Nomoto 2020
+:br2019-1: Data from Bravo, E. et al., Table 3
+:br2019-2: Data from Bravo, E. et al., Table 4
+:gro2021-1: Data from Gronow, S. et al., Tables 3/A10 He+Core detonations
+:gro2021-2: Data from Gronow, S. et al., Tables 4/A8 He+Core detonations
+:mor2018-1: Data from Mori, K. et al., W7
+:mor2018-2: Data from Mori, K. et al., WDD2
 
 Contributions
 =============
@@ -214,7 +214,7 @@ If you find a bug or have a question, please [open an issue in the project's rep
 
 Contributions are welcome, please read our `contributing guidelines`_.
 
-.. _`contributing guidelines`: https://github.com/xuanxu/starmatrix/blob/main/CONTRIBUTING.md
+.. _`contributing guidelines`: https://github.com/xuanxu/starmatrix/blob/review/CONTRIBUTING.md
 
 License
 =======
@@ -247,7 +247,7 @@ Starmatrix is built upon a long list of previous works from different authors/pa
 * *Leung & Nomoto*, 2018, ApJ, Vol 861, Issue 2, Id 143
 * *Leung & Nomoto*, 2020, ApJ, Vol 888, Issue 2, Id 80
 * *Bravo, E. et al.*, 2019, MNRAS, 482, Issue 4, 4346–4363
-* *Gronow, S. et al.*, 2021, A&A
+* *Gronow, S. et al.*, 2021, A&A 
 * *Mori, K. et al.*, 2018, ApJ, 863:176
 * *Chen, X., Hu, L. & Wang, L.*, 2021, ApJ
 * *Strolger et al*, 2020, ApJ, Vol 890, 2. doi: 10.3847/1538-4357/ab6a97
