### 2022-07-11 #4461 @ygrange
Mostly note to self than anything else: Looked at the replies and like them. Only thing I still want to do so:
- Have a look at the installation because I haven't checked it works now (though I have no reason not to believe it)
- Read the new version of the paper, and see if I have any comments left
Getting back at you when I did that.
