### 2022-07-11 #4461 @ygrange
installation works as documented now!

Few minor textual comments on the paper:
line 10: "This" -> "These" (I Think)
line 11: This is really minor, but since the sentence is quite long, but I still love that you explain in short what an SSP is and you already defined the acronym before, I would probably use SSP in stead of the written out form here.

That's it really. I will check the last box off my checklist and tell the editor that as far as I am concerned the review is done. Looking forward to this going public (And curious to see what the other reviewer has to see :) ).
