### 2022-07-12 #4461 @xuanxu
> Few minor textual comments on the paper:
> line 10: "This" -> "These" (I Think)
> line 11: This is really minor, but since the sentence is quite long, but I still love that you explain in short what an SSP is and you already defined the acronym before, I would probably use SSP in stead of the written out form here.

Changed!
