### 2022-07-15 xuanxu/starmatrix#59 @CFGrote
[JOSS Review]: Tests
# Testing as per README.md instructions after fresh install (into conda env)

## Fresh install
```
$> git pull --rebase
$> pip install -e ".[dev]"
...
Successfully installed attrs-21.4.0 coverage-6.4.2 iniconfig-1.1.1 packaging-21.3 pluggy-1.0.0 py-1.11.0 pyparsing-3.0.9 pytest-7.1.2 pytest-cov-3.0.0 pytest-mock-3.8.2 starmatrix-1.7.1 tomli-2.0.1
```
Also noted that `--cov=starmatrix` on the CL is redundant since already present in the setup.cfg/[tool:pytest line 66

```
$> pytest -v --cov=starmatrix                                                                                                                                                         [main]
==================================================================================== test session starts =====================================================================================
platform linux -- Python 3.10.4, pytest-7.1.2, pluggy-1.0.0 -- /opt/anaconda3/envs/starmatrix_review/bin/python
cachedir: .pytest_cache
rootdir: /home/XXX/Repositories/JOSSReviews/starmatrix, configfile: setup.cfg, testpaths: src/starmatrix/tests
plugins: cov-3.0.0, mock-3.8.2
collected 97 items

src/starmatrix/tests/test_abundances.py::test_select_abundances PASSED                                                                                                                 [  1%]
src/starmatrix/tests/test_abundances.py::test_abundances_classes_presence PASSED                                                                                                       [  2%]
src/starmatrix/tests/test_abundances.py::test_description_presence PASSED                                                                                                              [  3%]
src/starmatrix/tests/test_abundances.py::test_feh_with_z_zero PASSED                                                                                                                   [  4%]
src/starmatrix/tests/test_abundances.py::test_feh_with_z_non_zero PASSED                                                                                                               [  5%]
src/starmatrix/tests/test_abundances.py::test_h_he4_values PASSED                                                                                                                      [  6%]
src/starmatrix/tests/test_abundances.py::test_elements_value_presence PASSED                                                                                                           [  7%]
src/starmatrix/tests/test_abundances.py::test_abundance_values_presence PASSED                                                                                                         [  8%]
src/starmatrix/tests/test_abundances.py::test_cri_lim_exception PASSED                                                                                                                 [  9%]
src/starmatrix/tests/test_cli.py::test_option_generate_config PASSED                                                                                                                   [ 10%]
src/starmatrix/tests/test_cli.py::test_option_config PASSED                                                                                                                            [ 11%]
src/starmatrix/tests/test_cli.py::test_model_is_configured_properly PASSED                                                                                                             [ 12%]
src/starmatrix/tests/test_cli.py::test_model_is_run PASSED                                                                                                                             [ 13%]
src/starmatrix/tests/test_cli.py::test_creation_of_output_directory FAILED                                                                                                             [ 14%]
src/starmatrix/tests/test_cli.py::test_existent_output_directory PASSED                                                                                                                [ 15%]
src/starmatrix/tests/test_dtds.py::test_dtds_presence PASSED                                                                                                                           [ 16%]
src/starmatrix/tests/test_dtds.py::test_select_dtd PASSED                                                                                                                              [ 17%]
src/starmatrix/tests/test_dtds.py::test_no_negative_time_values PASSED                                                                                                                 [ 18%]
src/starmatrix/tests/test_dtds.py::test_dtd_correction_factor PASSED                                                                                                                   [ 19%]
src/starmatrix/tests/test_elements.py::test_read_file_on_init PASSED                                                                                                                   [ 20%]
src/starmatrix/tests/test_elements.py::test_interpolation_for_mass_data PASSED                                                                                                         [ 21%]
src/starmatrix/tests/test_elements.py::test_extrapolation_for_mass_data PASSED                                                                                                         [ 22%]
src/starmatrix/tests/test_elements.py::test_for_mass_with_no_yield_corrections PASSED                                                                                                  [ 23%]
src/starmatrix/tests/test_elements.py::test_for_mass_with_yield_corrections PASSED                                                                                                     [ 24%]
src/starmatrix/tests/test_elements.py::test_cri_lim_exception PASSED                                                                                                                   [ 25%]
src/starmatrix/tests/test_functions.py::test_value_in_interval PASSED                                                                                                                  [ 26%]
src/starmatrix/tests/test_functions.py::test_secondary_mass_fraction PASSED                                                                                                            [ 27%]
src/starmatrix/tests/test_functions.py::test_mean_lifetime_stellar_mass_relation PASSED                                                                                                [ 28%]
src/starmatrix/tests/test_functions.py::test_total_energy_no_negative_time_values PASSED                                                                                               [ 29%]
src/starmatrix/tests/test_functions.py::test_imf_zero PASSED                                                                                                                           [ 30%]
src/starmatrix/tests/test_functions.py::test_imf_binaries_are_zero_for_non_valid_masses PASSED                                                                                         [ 31%]
src/starmatrix/tests/test_functions.py::test_imf_binary_primary_integrates_phi_primary PASSED                                                                                          [ 32%]
src/starmatrix/tests/test_functions.py::test_imf_binary_secondary_integrates_phi_secondary PASSED                                                                                      [ 34%]
src/starmatrix/tests/test_functions.py::test_global_imf PASSED                                                                                                                         [ 35%]
src/starmatrix/tests/test_functions.py::test_imf_supernovae_II_non_zero_for_SNII_masses PASSED                                                                                         [ 36%]
src/starmatrix/tests/test_functions.py::test_imf_supernovae_II_is_zero_for_lower_masses PASSED                                                                                         [ 37%]
src/starmatrix/tests/test_functions.py::test_imf_supernovae_II_includes_binary_primaries PASSED                                                                                        [ 38%]
src/starmatrix/tests/test_functions.py::test_return_fractions PASSED                                                                                                                   [ 39%]
src/starmatrix/tests/test_imfs.py::test_select_imf PASSED                                                                                                                              [ 40%]
src/starmatrix/tests/test_imfs.py::test_valid_values_presence PASSED                                                                                                                   [ 41%]
src/starmatrix/tests/test_imfs.py::test_description_presence PASSED                                                                                                                    [ 42%]
src/starmatrix/tests/test_imfs.py::test_salpeter_alpha PASSED                                                                                                                          [ 43%]
src/starmatrix/tests/test_imfs.py::test_imf_is_zero_if_no_positive_mass PASSED                                                                                                         [ 44%]
src/starmatrix/tests/test_imfs.py::test_minimum_mass_value_for_kroupa_imfs PASSED                                                                                                      [ 45%]
src/starmatrix/tests/test_imfs.py::test_for_mass_is_normalized PASSED                                                                                                                  [ 46%]
src/starmatrix/tests/test_imfs.py::test_normalization_factor PASSED                                                                                                                    [ 47%]
src/starmatrix/tests/test_imfs.py::test_phi_m_phi_relation PASSED                                                                                                                      [ 48%]
src/starmatrix/tests/test_imfs.py::test_stars_per_mass_unit PASSED                                                                                                                     [ 49%]
src/starmatrix/tests/test_matrix.py::test_q_index PASSED                                                                                                                               [ 50%]
src/starmatrix/tests/test_matrix.py::test_empty_q_matrix PASSED                                                                                                                        [ 51%]
src/starmatrix/tests/test_matrix.py::test_matrices_are_empty_if_not_enough_mass PASSED                                                                                                 [ 52%]
src/starmatrix/tests/test_matrix.py::test_q_sn_size PASSED                                                                                                                             [ 53%]
src/starmatrix/tests/test_matrix.py::test_q_size PASSED                                                                                                                                [ 54%]
src/starmatrix/tests/test_matrix.py::test_q_applies_yield_corrections PASSED                                                                                                           [ 55%]
src/starmatrix/tests/test_matrix.py::test_q_with_no_yield_corrections PASSED                                                                                                           [ 56%]
src/starmatrix/tests/test_matrix.py::test_cri_lim_exception PASSED                                                                                                                     [ 57%]
src/starmatrix/tests/test_model.py::test_model_initialization PASSED                                                                                                                   [ 58%]
src/starmatrix/tests/test_model.py::test_model_run PASSED                                                                                                                              [ 59%]
src/starmatrix/tests/test_model.py::test_explosive_nucleosynthesis_with_logt_step PASSED                                                                                               [ 60%]
src/starmatrix/tests/test_model.py::test_explosive_nucleosynthesis_with_t_step PASSED                                                                                                  [ 61%]
src/starmatrix/tests/test_model.py::test_explosive_nucleosynthesis_with_two_steps_t PASSED                                                                                             [ 62%]
src/starmatrix/tests/test_model.py::test_explosive_nucleosynthesis_with_fixed_n_steps PASSED                                                                                           [ 63%]
src/starmatrix/tests/test_model.py::test_explosive_nucleosynthesis_with_invalid_step PASSED                                                                                            [ 64%]
src/starmatrix/tests/test_model.py::test_explosive_nucleosynthesis_step_t PASSED                                                                                                       [ 65%]
src/starmatrix/tests/test_model.py::test_explosive_nucleosynthesis_step_logt PASSED                                                                                                    [ 67%]
src/starmatrix/tests/test_model.py::test_explosive_nucleosynthesis_two_steps_t PASSED                                                                                                  [ 68%]
src/starmatrix/tests/test_model.py::test_explosive_nucleosynthesis_fixed_n_steps PASSED                                                                                                [ 69%]
src/starmatrix/tests/test_model.py::test_create_q_matrices PASSED                                                                                                                      [ 70%]
src/starmatrix/tests/test_model.py::test_create_q_matrices_all_zeros_files_if_wrong_data PASSED                                                                                        [ 71%]
src/starmatrix/tests/test_model.py::test_matrix_header PASSED                                                                                                                          [ 72%]
src/starmatrix/tests/test_model.py::test_return_fractions PASSED                                                                                                                       [ 73%]
src/starmatrix/tests/test_settings.py::test_defaults PASSED                                                                                                                            [ 74%]
src/starmatrix/tests/test_settings.py::test_default_settings_with_no_params PASSED                                                                                                     [ 75%]
src/starmatrix/tests/test_settings.py::test_default_settings_adds_extra_params PASSED                                                                                                  [ 76%]
src/starmatrix/tests/test_settings.py::test_default_settings_adds_extra_params_only_if_needed PASSED                                                                                   [ 77%]
src/starmatrix/tests/test_settings.py::test_validate_with_valid_values PASSED                                                                                                          [ 78%]
src/starmatrix/tests/test_settings.py::test_validate_with_invalid_values PASSED                                                                                                        [ 79%]
src/starmatrix/tests/test_settings.py::test_validate_max_mass_with_valid_values PASSED                                                                                                 [ 80%]
src/starmatrix/tests/test_settings.py::test_validate_max_mass_with_invalid_values PASSED                                                                                               [ 81%]
src/starmatrix/tests/test_settings.py::test_validate_mass_limits_for_starburst_imf PASSED                                                                                              [ 82%]
src/starmatrix/tests/test_settings.py::test_validate_with_invalid_settings PASSED                                                                                                      [ 83%]
src/starmatrix/tests/test_settings.py::test_validate_with_yield_corrections PASSED                                                                                                     [ 84%]
src/starmatrix/tests/test_settings.py::test_validate_yield_corrections_with_invalid_input PASSED                                                                                       [ 85%]
src/starmatrix/tests/test_settings.py::test_validate_yield_corrections PASSED                                                                                                          [ 86%]
src/starmatrix/tests/test_settings.py::test_deprecation_warnings PASSED                                                                                                                [ 87%]
src/starmatrix/tests/test_supernovae.py::test_empty_yields_set_is_all_zeroes PASSED                                                                                                    [ 88%]
src/starmatrix/tests/test_supernovae.py::test_yields_structure PASSED                                                                                                                  [ 89%]
src/starmatrix/tests/test_supernovae.py::test_yields_from_iwamoto PASSED                                                                                                               [ 90%]
src/starmatrix/tests/test_supernovae.py::test_yields_from_seitenzahl PASSED                                                                                                            [ 91%]
src/starmatrix/tests/test_supernovae.py::test_yields_from_leung_nomoto_2018 PASSED                                                                                                     [ 92%]
src/starmatrix/tests/test_supernovae.py::test_yields_from_leung_nomoto_2020 PASSED                                                                                                     [ 93%]
src/starmatrix/tests/test_supernovae.py::test_yields_from_bravo_2019 PASSED                                                                                                            [ 94%]
src/starmatrix/tests/test_supernovae.py::test_yields_from_gronow_2021 PASSED                                                                                                           [ 95%]
src/starmatrix/tests/dtds/test_strolger.py::test_parameters_initialization PASSED                                                                                                      [ 96%]
src/starmatrix/tests/dtds/test_strolger.py::test_has_description PASSED                                                                                                                [ 97%]
src/starmatrix/tests/dtds/test_strolger.py::test_is_normalized PASSED                                                                                                                  [ 98%]
src/starmatrix/tests/dtds/test_strolger.py::test_normalization_rate_uses_hubble_efficiency PASSED                                                                                      [100%]

========================================================================================== FAILURES ==========================================================================================
_____________________________________________________________________________ test_creation_of_output_directory ______________________________________________________________________________

[…]
