### 2022-07-15 xuanxu/starmatrix#60 @CFGrote
[JOSS Review] API documentation
I find the API documentation provided at https://starmatrix.readthedocs.io a bit rudimentary. I'd suggest to document classes and methods (along with  signature and return type) and standalone functions. You're already using docstrings, if you add
```
:param <parameter_name>: <description>
:type   <parameter_name>: <type>
```
to class methods and then add the  [sphinx autodoc extension](https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html), to the list of extensions in docs/conf.py, you'll have it. This would also activate online documentation in ipython or jupyter interactive environments.
