### 2022-07-15 xuanxu/starmatrix#61 @CFGrote
[JOSS Review] Interactive usage
I'm following the "Advanced" example at https://starmatrix.readthedocs.io/en/latest/usage.html#examples.

Now I'm at the step to define my custom parameters. I could set it all up one by one but what would really be helpful is to be able
to get a handle to the default variables and only customize the ones I want to vary. That's possible via `settings.default_settings()` but that is not mentioned in the tutorial. I suggest adding a corresponding line o the docs.

The rest of the example works just fine.
