### 2022-07-15 xuanxu/starmatrix#62 @CFGrote
[JOSS Review]: Units
I assume that all quantities are expressed in solar mass units and other domain specific unit systems. It should be mentioned clearly in
the code, README and paper.

I could imagine that this code be used as a pre or postprocessor to other codes that may employ a different system of units, so unit conversion would have to be considered. To ease such a use case, I would suggest to employ a physical units library such as pint. I understand this is a huge change to the codebase, therefore I leave this as recommendation for future work.
