### 2022-07-15 xuanxu/starmatrix#63 @CFGrote
[JOSS Review]: matrix datastructure
I suggest to return the Q matrix not as a plain array but as an [xarray](https://xarray.dev/) with appropriately labeled axes and dimensions. Otherwise every user will have to code up a index-to-element mapping which is a potential source of error that could easily be avoided. Getting the raw array out of an xarray is trivial.
