### 2022-07-15 #4461 @CFGrote
Raised issues on:
- [ ] Tests https://github.com/xuanxu/starmatrix/issues/59
- [ ] API documentation https://github.com/xuanxu/starmatrix/issues/60
- [ ] Interactive usage https://github.com/xuanxu/starmatrix/issues/61
- [ ] Physical units https://github.com/xuanxu/starmatrix/issues/62
- [ ] Q matrix datatype https://github.com/xuanxu/starmatrix/issues/63
