### 2022-07-15 #4461 @CFGrote
Just two minor comments on the paper, see attached.
[10.21105.joss.04461_annotated_CFGrote.pdf](https://github.com/openjournals/joss-reviews/files/9120150/10.21105.joss.04461_annotated_CFGrote.pdf)
