### 2022-07-15 xuanxu/starmatrix#61 @CFGrote
For what it's worth, I attach the notebook here

```python
import starmatrix
from starmatrix import settings
from starmatrix.model import Model
```

```python
custom_params = settings.default_settings()
```

```python
custom_params
```

    {'z': 0.02,
     'sol_ab': 'as09',
     'imf': 'kroupa2002',
     'imf_alpha': 2.35,
     'imf_m_low': 0.15,
     'imf_m_up': 100,
     'm_min': 0.98,
     'm_max': 40.0,
     'binary_fraction': 0.15,
     'dtd_sn': 'rlp',
     'dtd_correction_factor': 1.0,
     'sn_yields': 'iwa1998',
     'output_dir': 'results',
     'matrix_headers': True,
     'return_fractions': False,
     'integration_step': 'logt',
     'deprecation_warnings': True,
     'expelled_elements_filename': '/home/grotec/Repositories/JOSSReviews/starmatrix/src/starmatrix/sample_input/expelled_elements',
     'yield_corrections': {},
     'total_time_steps': 300}

```python
custom_params['binary_fraction'] = 0.13
custom_params['sol_ab'] = 'he10'
```

```python
context = settings.validate(custom_params)
```

```python
context
```

    {'z': 0.02,
     'sol_ab': 'he10',
     'imf': 'kroupa2002',
     'imf_alpha': 2.35,
     'imf_m_low': 0.15,
     'imf_m_up': 100,
     'm_min': 0.98,
     'm_max': 40.0,
     'binary_fraction': 0.13,
     'dtd_sn': 'rlp',
     'dtd_correction_factor': 1.0,
     'sn_yields': 'iwa1998',
     'output_dir': 'results',
     'matrix_headers': True,
     'return_fractions': False,
     'integration_step': 'logt',
     'deprecation_warnings': True,
     'expelled_elements_filename': '/home/grotec/Repositories/JOSSReviews/starmatrix/src/starmatrix/sample_input/expelled_elements',
     'total_time_steps': 300}

```python
Model(context).run()
```

```python
from starmatrix import functions as sm
from starmatrix import matrix
```

```python
stellar_mass = 5.09
z = 0.04
stellar_tau = sm.stellar_lifetime(stellar_mass, z)
```

```python
stellar_tau
```

    0.08900492008580678

```python
contribution_matrix = matrix.q_sn(stellar_mass)
```

```python
contribution_matrix
```

    array([[0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00],
           [0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00],
           [0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00],
           [0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00],
           [9.58504495e-03, 9.58504495e-03, 9.58504495e-03, 9.58504495e-03,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00],
           [2.83780834e-02, 2.83780834e-02, 2.83780834e-02, 2.83780834e-02,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00],
           [2.30199837e-07, 2.30199837e-07, 2.30199837e-07, 2.30199837e-07,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00],
           [2.77827390e-07, 2.77827390e-07, 2.77827390e-07, 2.77827390e-07,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00],
           [0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            2.11851191e-01],
           [4.00865234e-04, 4.00865234e-04, 4.00865234e-04, 4.00865234e-04,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00],
           [1.68680915e-03, 1.68680915e-03, 1.68680915e-03, 1.68680915e-03,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00],
           [3.05610129e-02, 3.05610129e-02, 3.05610129e-02, 3.05610129e-02,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00],
           [1.67887123e-02, 1.67887123e-02, 1.67887123e-02, 1.67887123e-02,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00],
           [2.36153281e-03, 2.36153281e-03, 2.36153281e-03, 2.36153281e-03,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00],
           [1.24228533e-01, 1.24228533e-01, 1.24228533e-01, 1.24228533e-01,
            0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
            0.00000000e+00]])

```python
contribution_matrix.shape
```

    (15, 9)

```python
starmatrix.constants.Q_MATRIX_ROWS, starmatrix.constants.Q_MATRIX_COLUMNS
```

    (15, 9)

```python
from matplotlib import pyplot as plt
```

```python

plt.imshow(contribution_matrix)
```
![output_15_1](https://user-images.githubusercontent.com/15476807/179209266-c8118a46-c1e1-4549-b4b6-c70fdff9779d.png)
