### 2022-07-15 #4461 @danielskatz
FYI, for future JOSS reviews, you can also suggest changes in the paper via a PR on the paper source, if that's easier
