+++ src/starmatrix/tests/test_cli.py
@@ -6,6 +6,7 @@ from pytest_mock import mocker
 import starmatrix.cli as cli
 import starmatrix.model as model
 import starmatrix.settings as settings
+import numpy.random as npr
 
 
 @pytest.fixture
@@ -29,11 +30,12 @@ def mock_config_file(mocker):
     """
     Fixture mocking contents of a config file reading
     """
-    config_file_content = "m_max: 33.0\ntotal_time_steps: 123"
+    random_output_dirname = "results-" + str(npr.default_rng().integers(low=9999, high=99999999))
+    config_file_content = "m_max: 33.0\ntotal_time_steps: 123\noutput_dir: " + random_output_dirname
     mocked_file = mocker.mock_open(read_data=config_file_content)
     mocker.patch.object(cli, 'open', mocked_file)
     mocker.spy(cli, "read_config_file")
-    return {'m_max': 33.0, 'total_time_steps': 123}
+    return {'m_max': 33.0, 'total_time_steps': 123, 'output_dir': random_output_dirname}
 
 
 @pytest.fixture
@@ -58,8 +60,8 @@ def test_option_config(mocker, deactivate_os_actions, mock_config_file):
 def test_model_is_configured_properly(mocker, deactivate_os_actions, mock_config_file):
     argparse.ArgumentParser.parse_args.return_value = argparse.Namespace(generate_config=False, config='ejectas.dat')
     cli.main()
-    expected_context = settings.validate(mock_config_file)
 
+    expected_context = settings.validate(mock_config_file)
     model.Model.assert_called_once_with(expected_context)
     model.Model(expected_context).run.assert_called()
 
@@ -71,11 +73,18 @@ def test_model_is_run(mocker, deactivate_os_actions):
 
 
 def test_creation_of_output_directory(mocker, deactivate_os_actions, mock_config_file):
+    argparse.ArgumentParser.parse_args.return_value = argparse.Namespace(generate_config=False, config='custom.conf')
     mocker.spy(cli, "create_output_directory")
     cli.main()
-    cli.create_output_directory.assert_called_once_with('results')
-    os.makedirs.assert_called_once_with('results')
+
+    expected_context = settings.validate(mock_config_file)
+    expected_output_dir = expected_context["output_dir"]
+
+    cli.create_output_directory.assert_called_once_with(expected_output_dir)
+    shutil.rmtree.assert_called_once_with(expected_output_dir, ignore_errors=True)
+    os.makedirs.assert_called_once_with(expected_output_dir)
     model.Model.assert_called()
+    model.Model(expected_context).run.assert_called()
 
 
 def test_existent_output_directory(mocker, deactivate_os_actions, mock_config_file, output_dir_existence):
