### 2022-07-15 4ffb684
Remove redundant options already present in the setup.cfg file
