### 2022-07-15 xuanxu/starmatrix#59 @xuanxu
Nice edge case: This failed because there existed a 'results' dir when running the tests. I've made the test [more robust](https://github.com/xuanxu/starmatrix/commit/1936428439c4043e7c89cdb106a4820e70d1055f) and now the error should not happen.

(redundant `-v --cov` removed from the docs)
