### 2022-07-15 cfcf282
Mention how to obtain the default settings
