### 2022-07-15 xuanxu/starmatrix#61 @xuanxu
Very nice, thanks!

[Docs](https://starmatrix.readthedocs.io/en/latest/usage.html) [updated](https://github.com/xuanxu/starmatrix/commit/cfcf28222e39fb8f5b02e49d5cfdaf875cf3cafc)!
