### 2022-07-15 b194b74
Mention the reference unit used for stellar masses
