### 2022-07-15 xuanxu/starmatrix#60 @xuanxu
Yes, this is a great suggestion, and was recommended also by @ygrange during his review. As [I mentioned there](https://github.com/xuanxu/starmatrix/issues/55#issuecomment-1176123328) I'm short on bandwith right now and when I'll work on it depends on how blocking you consider it for the current paper submission.
