### 2022-07-15 xuanxu/starmatrix#63 @xuanxu
We currently have this in our roadmap, we plan to add a setting to specify the desired format for the returned data (other than the current Numpy arrays), with hdf5 and xarray among the available options.

So this is something that should be ready for the 2.0 release, unless you think it is needed for the current version submitted to JOSS, in that case and about a timetible for it I can only refer to my [previous comment](https://github.com/xuanxu/starmatrix/issues/60#issuecomment-1185795311) 😓
