### 2022-07-16 xuanxu/starmatrix#62 @CFGrote
Ok, I'll close here and look forward to units support in 2.0 ;)
