### 2022-07-16 xuanxu/starmatrix#60 @CFGrote
Ok I'll close here as #55 addresses the same points.
