+++ paper/paper.md
@@ -31,21 +31,21 @@ Using explicit (and configurable) values for *solar abundances*, *metallicity (z
 
 # Statement of need
 
-Calculating the contributions of chemical elements from different stars in several mass ranges is a necessary step for galactic chemical evolution models [@molla2017]. Usually this is done internally and the chemical yields are convoluted with a star formation history (SFH) inside the model, making it difficult to find reusable opensourced code for calculating the yield for a simple stellar population. The central module in Chempy [@rybizki2017] and the SYGMA module from the NUPYCEE framework [@sygma2018] are exceptions, but being part of bigger codebases they are not available as standalone libraries, not listed in the official Python Package Index and learning how to install and use them is not simple.
+Calculating the contributions of chemical elements from different stars in several mass ranges is a necessary step for galactic chemical evolution models [@molla2017]. Usually this is done internally and the chemical yields are convoluted with a star formation history (SFH) inside the model, making it difficult to find reusable open-sourced code for calculating the yield for a simple stellar population. The central module in Chempy [@rybizki2017] and the SYGMA module from the NUPYCEE framework [@sygma2018] are exceptions, but being part of bigger codebases, they are not available as standalone libraries, not listed in the official Python Package Index, and learning how to install and use them is not simple.
 
-`Starmatrix` isolates the SSP yield calculation step as an open source permissively licensed Python implementation, in a modular and flexible way that can be used to provide detailed datasets to be used as input for galactic chemical evolution models, to compare the validity of different yield sets from the literature or to assess different nucleosynthesis modeling assumptions [@bazan2003].
+`Starmatrix` isolates the SSP yield calculation step as an open-source permissively-licensed Python implementation, in a modular and flexible way that can be used to provide detailed datasets to be used as input for galactic chemical evolution models, to compare the validity of different yield sets from the literature or to assess different nucleosynthesis modeling assumptions [@bazan2003].
 
 # Features
 
 In order to calculate $Q$ matrices for the whole range of selected stellar masses, the code estimates the stellar lifetime for the minimum and maximum masses and obtains a time interval that is then divided into a fixed number of steps to integrate in. The total time steps can be configured but also the integration step can be forced via settings to be constant in $t$ or in $log(t)$. For each time interval of the integration, the supernova rates are obtained using the corresponding IMF for core-collapse supernovae and a configurable delay time distribution for type Ia supernovae.
 Then, the time step is converted back into stellar mass intervals and a $Q$ matrix is calculated for that mass interval weighing it by the selected IMF to create the final output data.
 
-All quantities in the $Q$ matrix are calculated on the base of the ejected mass from stars, using for it the dataset of yields per stellar mass that is entered as input for the code. The reference unit used for stellar mass in the solar mass. The code is prepared to include any sets of stellar yields from the literature but if none is declared a default dataset for solar metallicity and with low and intermediate mass star yields from [@gavilan2006] and yields of massive stars from [@chieffi2004] will be used.
+All quantities in the $Q$ matrix are calculated on the basis of the ejected mass from stars, using for it the dataset of yields per stellar mass that is entered as input for the code. The reference unit used for stellar mass is the solar mass. The code is prepared to include any sets of stellar yields from the literature but if none is declared, a default dataset for solar metallicity and with low and intermediate mass star yields from @gavilan2006 and yields of massive stars from @chieffi2004 will be used.
 
-![A sample plot using the output data from several `Starmatrix` runs. Lines show true yields `p` of Oxygen using different datasets for low and intermediate mass yields (columns) and for massive stars yields (rows), for a range of metallicity values (bottom horizontal axis) and different IMFs available as Starmatix' settings.\label{fig:oxygen}](sample_plot.png)
[…]
