+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 `Starmatrix` is a Python package for computing the chemical contribution to the interstellar medium ejected by simple stellar populations (SSPs).
 
-One of the key ingredients of galactic chemical evolution (GCE) models is the nucleosynthetic contribution returned to the interstellar medium by the evolving stellar populations and supernovae. These yields vary depending on the age and metallicity of the stars and are combined following the mass distribution of stars to represent a SSP, stars formed at the same time and having the same initial element composition. Repeating this process adjusting for the the star formation at each time step is one of the main mechanisms of GCE models.
+One of the key ingredients of galactic chemical evolution (GCE) models is the nucleosynthetic contribution returned to the interstellar medium by the evolving stellar populations and supernovae. These yields vary depending on the age and metallicity of the stars and are combined following the mass distribution of stars to represent a SSP, stars formed at the same time and having the same initial element composition. Repeating this process, adjusting for the the star formation at each time step is one of the main mechanisms of GCE models.
 
 `Starmatrix` reads a single configuration file and calculates the combined contributions of chemical elements ejected to the interstellar medium during the lifetime of stars in the provided mass range. For each mass step an ejections matrix is computed for these fifteen elements: $H$, $D$, $^{3}He$, $^{4}He$, $^{12}C$, $^{16}O$, $^{14}N$, $^{13}C$, $^{20}Ne$, $^{24}Mg$, $^{28}Si$, $^{32}S$, $^{40}Ca$, $^{56}Fe$, and also all neutron-rich CNO isotopes as only one group.
 
@@ -37,7 +37,7 @@ Calculating the contributions of chemical elements from different stars in sever
 
 # Features
 
-In order to calculate $Q$ matrices for the whole range of selected stellar masses, the code estimates the stellar lifetime for the minimum and maximum masses and obtains a time interval that is then divided into a fixed number of steps to integrate in. The total time steps can be configured but also the integration step can be forced via settings to be constant in $t$ or in $log(t)$. For each time interval of the integration, the supernova rates are obtained using the corresponding IMF for core-collapse supernovae and a configurable delay time distribution for type Ia supernovae.
+In order to calculate $Q$ matrices for the whole range of selected stellar masses, the code estimates the stellar lifetime for the minimum and maximum masses and obtains a time interval that is then divided into a fixed number of steps to integrate. The total time steps can be configured but also the integration step can be forced via settings to be constant in $t$ or in $log(t)$. For each time interval of the integration, the supernova rates are obtained using the corresponding IMF for core-collapse supernovae and a configurable delay time distribution for type Ia supernovae.
 Then, the time step is converted back into stellar mass intervals and a $Q$ matrix is calculated for that mass interval weighing it by the selected IMF to create the final output data.
 
 All quantities in the $Q$ matrix are calculated on the basis of the ejected mass from stars, using for it the dataset of yields per stellar mass that is entered as input for the code. The reference unit used for stellar mass is the solar mass. The code is prepared to include any sets of stellar yields from the literature but if none is declared, a default dataset for solar metallicity and with low and intermediate mass star yields from @gavilan2006 and yields of massive stars from @chieffi2004 will be used.
