### 2022-07-19 #4461 @xuanxu
> * I don't quite get "Repeating this process adjusting for the the star formation at each time step" - should this be "Repeating this process, adjusting for the the star formation at each time step," ?  Or something else?

Updated!

> * I think the "as" in "masses of elements j ejected to the galactic medium as element i" doesn't quite make sense.

This is literally the amount of an element that is converted by nucleosynthesis into another element, so the data in the matrix is (for instance) the mass that initially is present in the star as Hydrogen but it is ejected as Oxygen.

> * Should "divided into a fixed number of steps to integrate in" be "divided into a fixed number of steps to integrate"?

Changed!
