### 2022-07-19 #4461 @danielskatz
> > * I think the "as" in "masses of elements j ejected to the galactic medium as element i" doesn't quite make sense.
>
> This is literally the amount of an element that is converted by nucleosynthesis into another element, so the data in the matrix is (for instance) the mass that initially is present in the star as Hydrogen but it is ejected as Oxygen.

Then maybe "masses of elements j converted to element in and ejected to the galactic medium"?
