### 2022-07-19 #4461 @danielskatz
Also, see https://github.com/xuanxu/starmatrix/pull/66 related to the first suggestion
